FossilOrigin-Name: 31e8d8ba9a1ed9fa78b8e7b29ef1811e5d09f458e9c16f244d9773189825d92b

+++ benchmark_card_data_import.py
@@ -25,6 +25,7 @@ import mtg_proxy_printer.model.carddb
 class Namespace:
     database_path: Path
     card_data: Path
+    keep: bool
 
 
 def parse_args() -> Namespace:
@@ -35,18 +36,22 @@ def parse_args() -> Namespace:
     parser.add_argument(
         "card_data", type=Path,
         help="'All cards' bulk data export from the Scryfall API. May be plain-text JSON or GZIP compressed JSON.")
+    parser.add_argument(
+        "-k", "--keep", action="store_true",
+        help="Re-use an existing database, performing an in-place card data update, instead of populating an empty, new database.")
     return parser.parse_args()
 
 
 to_be_profiled_functions = {
     mtg_proxy_printer.card_info_downloader.CardInfoDownloadWorker: [
         "populate_database",
-        "read_json_card_data",
     ],
     mtg_proxy_printer.card_info_downloader: [
         "_insert_set",
         "_insert_card_faces",
         "_should_skip_card",
+        "_clean_unused_data",
+        "_remove_card"
     ],
     # Bypass the lru_cache
     mtg_proxy_printer.card_info_downloader._insert_card: [
@@ -91,10 +96,13 @@ if __name__ == "__main__":
         print("Running with kernprof. Injecting profile decorator.")
         inject_line_profiler()
     args = parse_args()
-    if args.database_path.exists():
+    if args.database_path.exists() and not args.keep:
         print("Deleting existing database…")
         args.database_path.unlink()
-    print("Creating new database…")
+    if not args.database_path.exists():
+        print("Creating new database…")
+    elif args.keep:
+        print("Re-use existing database…")
     cdb = mtg_proxy_printer.model.carddb.CardDatabase(args.database_path)
     cid = mtg_proxy_printer.card_info_downloader.CardInfoDownloadWorker(cdb)
     print("Starting benchmark…")
